llama-index-starter-tutorial/requirements.txt
llama_index
black
python-dotenv
llama-index-llms-ollama
llama-index-embeddings-huggingface